handle app ready state in store
