+++ dash-renderer/src/actions/constants.js
@@ -9,6 +9,7 @@ const actionList = {
     ON_ERROR: 'ON_ERROR',
     RESOLVE_ERROR: 'RESOLVE_ERROR',
     SET_HOOKS: 'SET_HOOKS',
+    SET_APP_READY: 'SET_APP_READY',
 };
 
 export const getAction = action => {
