+++ dash-renderer/src/reducers/isAppReady.js
@@ -1,0 +1,8 @@
+import {getAction} from '../actions/constants';
+
+export default function config(state = false, action) {
+    if (action.type === getAction('SET_APP_READY')) {
+        return action.payload;
+    }
+    return state;
+}
